requirements.txt
numpy
pynrrd
tifffile
scikit-image
matplotlib
networkx
scipy
torch
opencv-python
scikit-learn